vecmat: Remove unused/incorrect xxx_quick_mag() routines
The routine vm_vec_normalize_quick_mag() was incorrect. Instead of
dividing the vector elements by the approximate magnitude, it was
multiplying each element by the approximate magnitude. This routine
was only used by vm_vec_normalized_dir_quick_mag() which fortunately
was not used anywhere.

Given the above, it was thought that removing both routines was the
correct course of action.

Signed-off-by: Peter Mitsis <peter.mitsis@gmail.com>